+++ src/main/java/pl/jiraput/repository/CompanyRepository.java
@@ -5,8 +5,10 @@ import org.springframework.stereotype.Repository;
 
 import pl.jiraput.model.Company;
 
+import java.math.BigInteger;
+
 @Repository
 public interface CompanyRepository extends JpaRepository<Company, Integer> {
-	Company findByTaxNumber(Integer taxNumber);
+	Company findByTaxNumber(BigInteger taxNumber);
 	Company findByName(String name);
 }
